LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/stylesheet-title-002-expected.txt


FAIL Title attribute in stylesheets not in the document tree is ignored assert_equals: expected 3 but got 2
